requirements.txt
async-timeout==4.0.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
fastavro==1.7.3
idna==3.4
jsonpath-ng==1.5.3
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==2.0.0
pbr==5.11.1
plux==1.3.1
ply==3.11
protobuf==3.20.3
pulsar-client==3.0.0
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
schedule==1.2.0
singleton-decorator==1.0.0
six==1.16.0
stevedore==5.0.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
